+++ src/app/core/models/enum/enum.model.ts
@@ -421,7 +421,7 @@ export enum QBDOnboardingState {
     CATEGORY_MAPPING = 'CATEGORY_MAPPING'
   }
 
-  export enum AccountingExportCreationType{
+  export enum AccountingExport{
     CREATING_DIRECT_COST = 'CREATING_DIRECT_COST',
     CREATING_PURCHASE_INVOICE = 'CREATING_PURCHASE_INVOICE'
   }
@@ -455,4 +455,9 @@ export enum QBDOnboardingState {
 
  export enum IntegrationName {
   SAGE300 = 'SAGE300'
+ }
+
+ export enum FundSource {
+  PERSONAL = 'PERSONAL',
+  CCC = 'CCC'
  }
