+++ src/app/core/models/sage300/db/sage300-accounting-export.model.ts
@@ -1,7 +1,7 @@
-import { AccountingExportCreationType } from "../../db/accounting-export.model";
+import { AccountingExport } from "../../db/accounting-export.model";
 import { PaginatedResponse } from "../../db/paginated-response.model";
 
-export interface Sage300AccountingExport extends AccountingExportCreationType {
+export interface Sage300AccountingExport extends AccountingExport {
     sage_300_errors: any;
     purchase_invoice_id: null | number;
     direct_costs_id: null | number;
