+++ src/app/core/services/common/export-log.service.ts
@@ -4,9 +4,11 @@ import { UserService } from '../misc/user.service';
 import { WorkspaceService } from './workspace.service';
 import { environment } from 'src/environments/environment';
 import { ExpenseGroupDescription, SkipExportLogResponse } from '../../models/si/db/expense-group.model';
-import { FyleReferenceType } from '../../models/enum/enum.model';
+import { AccountingExportStatus, FyleReferenceType } from '../../models/enum/enum.model';
 import { Observable } from 'rxjs';
-import { Sage300AccountingExport } from '../../models/sage300/db/sage300-accounting-export.model';
+import { AccountingExportResponse, Sage300AccountingExport } from '../../models/sage300/db/sage300-accounting-export.model';
+import { AccountingExport } from '../../models/db/accounting-export.model';
+import { SelectedDateFilter } from '../../models/qbd/misc/date-filter.model';
 
 @Injectable({
   providedIn: 'root'
@@ -37,13 +39,13 @@ export class ExportLogService {
     return referenceType;
   }
 
-  getSkipExportLogs(limit: number, offset: number): Observable<SkipExportLogResponse> {
+  getSkippedExpenses(limit: number, offset: number): Observable<SkipExportLogResponse> {
     const workspaceId = this.workspaceService.getWorkspaceId();
 
     return this.apiService.get(`/workspaces/${workspaceId}/fyle/expenses/`, {limit, offset});
   }
 
-  generateFyleUrl(accountingExport: Sage300AccountingExport, referenceType: FyleReferenceType) : string {
+  generateFyleUrl(accountingExport: AccountingExport, referenceType: FyleReferenceType) : string {
     let url = `${environment.fyle_app_url}/app/`;
     if (referenceType === FyleReferenceType.EXPENSE) {
       url += `main/#/view_expense/${accountingExport.expenses[0].expense_id}`;
@@ -54,6 +56,28 @@ export class ExportLogService {
     }
 
     return `${url}?org_id=${this.org_id}`;
+  }
+
+
+  getAccountingExports(state: AccountingExportStatus | AccountingExportStatus.COMPLETE, limit: number, offset: number, selectedDateFilter: SelectedDateFilter | null, exportedAt?: Date | null): Observable<AccountingExportResponse> {
+    const params: any = {
+      limit,
+      offset
+    };
+    params.state = state;
+
+    if (selectedDateFilter) {
+      const startDate = selectedDateFilter.startDate.toLocaleDateString().split('/');
+      const endDate = selectedDateFilter.endDate.toLocaleDateString().split('/');
+      params.start_date = `${startDate[2]}-${startDate[1]}-${startDate[0]}T00:00:00`;
+      params.end_date = `${endDate[2]}-${endDate[1]}-${endDate[0]}T23:59:59`;
+    }
+
+    if (exportedAt) {
+      params.exported_at = exportedAt;
+    }
+
+    return this.apiService.get(`/workspaces/${this.workspaceId}/fyle/accounting_export/`, params);
   }
 
 }
