+++ src/app/shared/components/export-log/export-log-table/export-log-table.component.ts
@@ -1,8 +1,6 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { AccountingExportCreationType, AccountingExportList } from 'src/app/core/models/db/accounting-export.model';
+import { AccountingExportList } from 'src/app/core/models/db/accounting-export.model';
 import { AppName } from 'src/app/core/models/enum/enum.model';
-import { ExpenseGroupList } from 'src/app/core/models/si/db/expense-group.model';
-import { Expense } from 'src/app/core/models/si/db/expense.model';
 import { WindowService } from 'src/app/core/services/common/window.service';
 
 @Component({
